Integrated pages and correct TS errors
