+++ src/app/(Organisation)/Dashboard/page.tsx
@@ -1,59 +1,8 @@
 import VisitorsChart from '@/components/organisation/VisitorsChart';
 import RevenueChart from '@/components/organisation/RevenueChart';
-import Image from 'next/image';
 import React from 'react'
-
-interface EntityCardProps {
-    title: string,
-    value: number,
-    imgPath: string
-}
-
-const EntityCard = ({title, value, imgPath}: EntityCardProps) => {
-    return (
-        <div className='bg-white rounded-md p-1 font-bold flex flex-row items-center gap-2 w-72 h-24 m-2 border border-primary-blue/20 hover:border-primary-blue/25 hover:shadow-sm'>
-            <div className='relative w-full h-full'>
-                {/* Fill parent div */}
-                <Image src={imgPath} alt="Car" fill style={{ 'objectFit': 'contain'}} />
-            </div>
-            <div className='w-1 h-3/5 bg-primary-blue/20'></div>
-            <div className='flex flex-col w-full pl-4'>
-                <h2 className='my-2'>{title}</h2>
-                <p className='text-2xl'>{value}</p>
-            </div>
-        </div>
-    );
-}
-
-interface HightlightCardProps {
-    title: string
-    topResources: { name: string, nbRequests: number, imgPath: string}[],
-}
-
-const HighlightCard = ({title, topResources}: HightlightCardProps) => {
-    return (
-        <div className='w-5/12 bg-white border-gray-200 shadow-md p-2 rounded-md'>
-            <div className='border-b-2 border-primary-blue/20 p-2 text-primary-blue text-bold text-xl'>
-                {title}
-            </div>
-            
-            {topResources.map((resource, index) => (
-                <div key={index} className='flex flex-row items-center gap-6 pl-6 mt-4'>
-                    <div className='flex items-center justify-center rounded-full border border-primary-blue/50 text-center w-10 h-10 '>{index + 1}</div>
-                    <div className='flex flex-row gap-2'>
-                        <div className='flex flex-row items-center gap-2'>
-                            <Image src={resource.imgPath} alt="Car" width={48} height={48} />
-                        </div>
-                        <div>
-                            <h2 className='font-bold'>{resource.name}</h2>
-                            <p className='text-gray-600'>{resource.nbRequests} Sollicitations</p>
-                        </div>
-                    </div> 
-                </div>
-            ))}
-        </div>
-    );
-}
+import HighlightCard from '@/components/organisation/HighlightCard';
+import EntityCard from '@/components/organisation/EntityCard';
 
 export default function DashboardPage() {
     const cardContents = [
@@ -108,7 +57,7 @@ export default function DashboardPage() {
                 </div>
                 <div className='w-full h-full'>
                     <div className=''>
-                        <RevenueChart revenues = {revenues} totalRevenue={1500000}/>
+                        <RevenueChart revenues = {revenues} />
                     </div>
                     <div className='w-full flex flex-row justify-center gap-8'>
                         <HighlightCard title='Top Drivers' topResources={topDrivers} />
