--- src/app/(Organisation)/cars/drivers.json
@@ -1,109 +1,0 @@
-{
-  "drivers": [
-    {
-      "id": 1,
-      "first_name": "John",
-      "last_name": "Doe",
-      "age": 35,
-      "license_number": "D1234567",
-      "license_type": "Class C",
-      "address": "123 Main St, City, CA 90001",
-      "phone": "+1 123-456-7890",
-      "email": "john.doe@example.com",
-      "vehicle_assigned": {
-        "id": 1,
-        "make": "Toyota",
-        "model": "Camry",
-        "year": 2020
-      },
-      "rating": 4.5,
-      "insurance_provider": "State Farm",
-      "insurance_policy": "SF123456789",
-      "profile_picture": "/assets/organisation/cars/1.jpeg"
-    },
-    {
-      "id": 2,
-      "first_name": "Jane",
-      "last_name": "Smith",
-      "age": 28,
-      "license_number": "S7654321",
-      "license_type": "Class M",
-      "address": "456 Park Ave, Miami, FL 33101",
-      "phone": "+1 987-654-3210",
-      "email": "jane.smith@example.com",
-      "vehicle_assigned": {
-        "id": 2,
-        "make": "Harley-Davidson",
-        "model": "Sportster",
-        "year": 2021
-      },
-      "rating": 4.8,
-      "insurance_provider": "Allstate",
-      "insurance_policy": "AL987654321",
-      "profile_picture": "/assets/organisation/cars/2.jpeg"
-    },
-    {
-      "id": 3,
-      "first_name": "Tom",
-      "last_name": "Harris",
-      "age": 42,
-      "license_number": "T12345678",
-      "license_type": "Class A",
-      "address": "789 Industrial Blvd, Dallas, TX 75201",
-      "phone": "+1 321-987-6543",
-      "email": "tom.harris@example.com",
-      "vehicle_assigned": {
-        "id": 3,
-        "make": "Ford",
-        "model": "F-150",
-        "year": 2019
-      },
-      "rating": 4.6,
-      "insurance_provider": "Progressive",
-      "insurance_policy": "PR1234567890",
-      "profile_picture": "/assets/organisation/cars/3.jpeg"
-    },
-    {
-      "id": 4,
-      "first_name": "Alice",
-      "last_name": "Johnson",
-      "age": 30,
-      "license_number": "J9876543",
-      "license_type": "Class B",
-      "address": "234 Oak St, New York, NY 10001",
-      "phone": "+1 432-987-6543",
-      "email": "alice.johnson@example.com",
-      "vehicle_assigned": {
-        "id": 4,
-        "make": "BMW",
-        "model": "X5",
-        "year": 2022
-      },
-      "rating": 4.7,
-      "insurance_provider": "Geico",
-      "insurance_policy": "GE123456789",
-      "profile_picture": "/assets/organisation/cars/4.jpeg"
-    },
-    {
-      "id": 5,
-      "first_name": "Mike",
-      "last_name": "Brown",
-      "age": 38,
-      "license_number": "M1239876",
-      "license_type": "Class C",
-      "address": "567 Maple Rd, Seattle, WA 98101",
-      "phone": "+1 555-123-4567",
-      "email": "mike.brown@example.com",
-      "vehicle_assigned": {
-        "id": 5,
-        "make": "Mercedes-Benz",
-        "model": "Metris",
-        "year": 2020
-      },
-      "rating": 4.9,
-      "insurance_provider": "Nationwide",
-      "insurance_policy": "NW123456789",
-      "profile_picture": "/assets/organisation/cars/5.jpeg"
-    }
-  ]
-}
