+++ src/utils/types/notifications.tsx
@@ -1,0 +1,8 @@
+export interface NotificationType {
+    id: number;
+    title: string;
+    content: string;
+    category: string;
+    created_at: string;
+    read_status: string;
+}
